+++ src/main/java/net/minecraftforge/gradle/tasks/Download.java
@@ -22,8 +22,6 @@ package net.minecraftforge.gradle.tasks;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
 import java.net.HttpURLConnection;
 import java.net.URL;